+++ CHANGELOG.md
@@ -3,8 +3,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
-### 🔄 Changed
-
+## StreamChatUI
+### 🐞 Fixed
+- Fix multiple messages staying highlighted when jumping to a message in a thread [#4216](https://github.com/GetStream/stream-chat-swift/pull/4216)
+  
 # [5.9.0](https://github.com/GetStream/stream-chat-swift/releases/tag/5.9.0)
 _August 12, 2026_